refactor(server): promote patchPath to shared/paths (BET-339)
The box server's Claude credential refresh (src/server/opencode.mjs)
hardcoded '/opt/homebrew/bin' literals next to the macOS PATH prefix.
The desktop plugin runner already factored that into patchPath in BET-327
(PR #284); the box server was deliberately left out for scope.

This PR promotes patchPath (and the DARWIN_PATH_PREFIX it depends on)
from src/main/capExecutor.ts into src/shared/paths.mjs so both
processes share one source of truth:

- src/shared/paths.mjs: adds DARWIN_PATH_PREFIX + patchPath (byte-identical
  body to the BET-327 implementation; renames nothing).
- src/shared/paths.d.mts: exposes the new exports with the same shape as
  the .mjs side.
- src/main/capExecutor.ts: imports patchPath from ../shared/paths.mjs,
  drops the local definition + DARWIN_PATH_PREFIX + the now-unused
  'delimiter' import. makeExec continues to call patchPath unchanged.
- src/main/capExecutor.test.ts: drops the patchPath describe block
  (relocated to paths.test.ts alongside the implementation).
- src/server/opencode.mjs: drops '/opt/homebrew/bin/claude' from
  resolveClaudeBin() candidates (the path patch below already prepends
  the macOS Homebrew prefix, so a Homebrew-installed claude is visible
  to the spawned child without an explicit probe). Replaces the hand-
  rolled '[..., process.env.PATH ?? ""].filter(Boolean).join(":")'
  augmentedPath with patchPath(baseEnv, process.platform). The
  '~/.local/bin' prefix stays in baseEnv — it is the resolver's concern,
  not the platform helper's.
- src/shared/paths.test.ts: adds linux-branch tests (no PATH key written
  when one is missing; byte-identical when one is present) plus a
  DARWIN_PATH_PREFIX content pin, alongside the relocated darwin/win32
  cases from the BET-327 test suite.

Acceptance:
- grep -rn '/opt/homebrew' src/server/ returns no hits.
- npm run typecheck && npm test green (964 pass, 0 fail).

Out of scope (per issue):
- desktop plugin runner is already patched (BET-327).
- tmuxSpawnEnv's LC_ALL handling — already platform-arg aware.
- spawnErrorMessage / killTree stay in capExecutor.ts.
